% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compute_post_inc_prob}
\alias{compute_post_inc_prob}
\title{Compute posterior inclusion probabilities}
\usage{
compute_post_inc_prob(
  mvbd_fit,
  risk_names = NULL,
  pred_names = NULL,
  digits = NULL
)
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{risk_names}{Names of the competing risks}

\item{pred_names}{Names of the predictors}

\item{digits}{If provided, the result is rounded to this many digits.}
}
\value{
Matrix of posterior inclusion probabilities with each row
 corresponding to a predictor and each column to a risk. The row names of the
 matrix are `pred_names`, if provided. The column names or `risk_names` or,
 if multiple risks, "Risk 1", etc.
}
\description{
Compute the posterior inclusion probabilities of the regression coefficients.
}
