% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_hazard_rate_compete}
\alias{log_hazard_rate_compete}
\title{Log of the hazard rate}
\usage{
log_hazard_rate_compete(alpha)
}
\arguments{
\item{alpha}{Vector of baseline hazard parameters}
}
\description{
Log of the hazard rate
}
\keyword{internal}
