% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_posterior_unaugmented}
\alias{log_posterior_unaugmented}
\title{Log of posterior for discrete-time survival with `n_risk` risks}
\usage{
log_posterior_unaugmented(
  n_vec,
  alpha_star,
  beta,
  X,
  time_to_event,
  event,
  cens,
  log_prior_n_vec,
  mu_alpha,
  s2_alpha
)
}
\description{
Here, the prior on cause-specific change points p(z_t | gamma_t) is not
included in `log_prior_n_vec`, and thus not included in this log-posterior
computation.
}
\keyword{internal}
