% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_prior_alpha}
\alias{log_prior_alpha}
\title{Log of the prior on alpha conditionally on the change points}
\arguments{
\item{alpha_star}{Matrix of unique values of alpha over each time interval}

\item{mu_alpha}{Mean of the marginal prior on alpha}

\item{s2_alpha}{Variance of the marginal prior on alpha}
}
\description{
Log of the prior on alpha conditionally on the change points
}
\keyword{internal}
