% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvb_detector.R
\name{log_prior_n_regimes}
\alias{log_prior_n_regimes}
\title{Log of the prior on `n_regimes` up to proportionality}
\usage{
log_prior_n_regimes(n_regimes)
}
\arguments{
\item{n_regimes}{No. of regimes, which is the no. of change points + 1}
}
\description{
This implements the Geometric distribution with success probability 0.5 used
 in the paper.
}
\keyword{internal}
