\name{mvb.detector-package}
\alias{mvb.detector-package}
\alias{mvb.detector}
\docType{package}
\title{\packageTitle{mvb.detector}}
\description{\packageDescription{mvb.detector}}
\section{Package Content}{\packageIndices{mvb.detector}}
\author{\packageAuthor{mvb.detector}}
\section{Maintainer}{\packageMaintainer{mvb.detector}}
\keyword{package}
