% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_baseline_hazards}
\alias{plot_baseline_hazards}
\title{Plot baseline hazards}
\usage{
plot_baseline_hazards(
  mvbd_fit,
  alpha_true = NULL,
  xlab = expression(paste("Time ", italic(t))),
  ylim = NULL,
  main_vec = NULL
)
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{alpha_true}{Matrix with true population parameters}

\item{xlab}{x-axis label}

\item{ylim}{Numeric vector of length 2, giving the y coordinates range}

\item{main_vec}{Plot titles}
}
\description{
Plot the posterior inference on the baseline hazards
}
