% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_change_points}
\alias{plot_change_points}
\title{Plot posterior inference on change points}
\usage{
plot_change_points(
  mvbd_fit,
  n_vec_true = NULL,
  z_star_true = NULL,
  BF = NULL,
  xlab = expression(paste("Time ", italic(t))),
  main_vec = NULL,
  BF_max = 100
)
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{n_vec_true}{Vector of true population regime durations}

\item{z_star_true}{Array specifying which risk are affected at change points}

\item{BF}{Whether to plot Bayes factors (TRUE), posterior probabilities
(FALSE) or both (NULL, default)}

\item{xlab}{x-axis label}

\item{main_vec}{Plot titles}

\item{BF_max}{Upper limit of the y-axis range for the Bayes factors}
}
\description{
Plot the posterior inference on the location of change points using Bayes
factors or posterior probabilities.
}
