% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_CHF}
\alias{plot_CHF}
\title{Plot cumulative hazard function}
\usage{
plot_CHF(mvbd_fit, xlab = expression(paste("Time ", italic(t))))
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{xlab}{x-axis label}
}
\description{
Plot the posterior inference on the cumulative hazard function for a subject
with all covariates equal to zero.
}
