% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_K}
\alias{plot_K}
\title{Plot posterior inference on number of change points}
\usage{
plot_K(mvbd_fit)
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}
}
\description{
Plot posterior distribution of the number of overall change points.
}
