% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_regression}
\alias{plot_regression}
\title{Plot posterior inference on regression coefficients}
\usage{
plot_regression(
  mvbd_fit,
  risk_names = NULL,
  pred_names = NULL,
  pred_names_offset = 0.07,
  inc_prob = FALSE,
  pred_group = NULL
)
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{risk_names}{Names of the competing risks for use in the legend}

\item{pred_names}{Names of the predictors shown next to the y-axis}

\item{pred_names_offset}{Vertical offset from the y-axis of the labels in
`pred_names`}

\item{inc_prob}{If true, the posterior inclusion probabilities are plot
instead of the posterior means and credible intervals.}

\item{pred_group}{Group indicator for predictors that should be jointly added
or removed when it comes to variable selection. Only used if
inc_prob = TRUE.}
}
\description{
Plot posterior means and 95\% credible intervals of the regression
coefficients.
}
