% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot MCMC trace plots}
\usage{
plot_trace(mvbd_fit, quantity = c("log_lik", "K"))
}
\arguments{
\item{mvbd_fit}{Output from the function \link{run_mvb_detector}}

\item{quantity}{Vector specifying which trace plots are generated}
}
\description{
Plot trace plots of the log-likelihood and/or the number of change points.
}
