% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvb_detector.R
\name{run_mvb_detector}
\alias{run_mvb_detector}
\title{Run the multivariate Bernoulli detector}
\usage{
run_mvb_detector(
  data,
  t_max = NULL,
  n_iter = 10000L,
  pred_group = NULL,
  burnin = n_iter\%/\%10L,
  show_progress_bar = TRUE,
  cp_constraint = TRUE
)
}
\arguments{
\item{data}{Data frame with as first variable an integer vector `time` with
the time-to-event data, as second variable a factor `event` of event
types/causes with the first level indicating that the time was censored, and
optionally additional variables that are treated as predictor values.}

\item{t_max}{The integer \eqn{t_{\max}}}

\item{n_iter}{Number of MCMC iterations, including burn-in}

\item{pred_group}{Group indicator for predictors that should be jointly added
or removed when it comes to variable selection}

\item{burnin}{Number of burn-in MCMC iterations}

\item{show_progress_bar}{Whether to show a progress bar}

\item{cp_constraint}{Whether to limit where change points are allowed in the
prior}
}
\value{
A list with the following items:
\tabular{ll}{
  \code{alpha_MCMC} \tab Array with MCMC samples of alpha \cr
  \code{beta_MCMC} \tab Array with MCMC samples of beta \cr
  \code{n_regimes_MCMC} \tab Vector with MCMC samples of the no. of regimes,
    which is the no. of change points + 1 \cr
  \code{log_lik_MCMC} \tab Vector with log-likelihood values of the MCMC
    samples. \cr
  \code{data} \tab The value of the function argument \code{data} \cr
  \code{data} \tab The value of the function argument \code{cp_constraint}
    \cr
}
}
\description{
Function that runs the MCMC algorithm for the multivariate Bernoulli detector
}
