% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_r_i_mat}
\alias{sample_r_i_mat}
\title{Parallel sampling of mixture indicators}
\usage{
sample_r_i_mat(U, z_min_eta, m_mix, tmp1, tmp2)
}
\description{
Parallel sampling of mixture indicators
}
\keyword{internal}
