% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvb_detector.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(n, alpha, X = NULL, beta = 0)
}
\arguments{
\item{n}{Number of observations}

\item{alpha}{Matrix with baseline hazard parameters}

\item{X}{Matrix with predictors}

\item{beta}{Matrix with cause-specific regression coefficients}
}
\value{
Data frame with an integer vector `time` with the time-to-event
 data, a factor `event` of event types/causes with the first level indicating
 that the time was censored, and optionally additional variables that are the
 predictor values provided by the argument `X`.
}
\description{
Simulate competing risks discrete survival data from the multinomial logit
model for the cause-specific hazard functions.
}
